added relPoints
Added a feature so that curve points can be specified relative to previus points. Add "+" to the beginning of a point position coordinate to make it relative to the previous one (.e.g "+-5" will be -5 units back from the previous point). Or use relPoints="true" at the entity attributes to make point number 2+ relative. This is just like the SVG relative positions.

Finally, added a feature so that if a point position has just two values they are the X and Y coordinates, and Z is automatically set to 0.